Add functions to module js file
