+++ src/modules/functions.js
@@ -1,5 +1,6 @@
 import menuImgSrc from '../menu.png';
-
+import enterImgSrc from '../enter.png';
+import refreshImgSrc from '../refresh.png';
 /* const addText = () => {
   const t = document.createElement('div');
   const text = "text's test";
@@ -13,7 +14,31 @@ import menuImgSrc from '../menu.png';
 };
 export default addText; */
 
-const addLocalStorage = () => {
+export const createFrame = () => {
+  const h1Container = document.querySelector('.h1-container');
+  const refreshImg = document.createElement('img');
+  const container = document.querySelector('.to-do-list');
+  refreshImg.src = refreshImgSrc;
+  refreshImg.className = 'refresh-img';
+  h1Container.appendChild(refreshImg);
+  const form = document.createElement('form');
+  const input = document.createElement('input');
+  input.type = 'text';
+  input.name = 'description';
+  input.placeholder = 'Add to your list...';
+  const inputLine = document.createElement('div');
+  inputLine.className = 'input-line';
+  inputLine.appendChild(input);
+  const enterImg = document.createElement('img');
+  enterImg.src = enterImgSrc;
+  enterImg.className = 'enter-img';
+
+  inputLine.appendChild(enterImg);
+  form.appendChild(inputLine);
+  container.appendChild(form);
+};
+
+export const addLocalStorage = () => {
   const ul = document.createElement('ul');
   const arrayOfTasks = [];
   const storedTasks = JSON.parse(localStorage.getItem('tasks'));
@@ -40,4 +65,13 @@ const addLocalStorage = () => {
   }
   return { ul, arrayOfTasks };
 };
-export default addLocalStorage;
+
+export const Task = {
+  create(description, index) {
+    const task = Object.create(this);
+    task.description = description;
+    task.completed = false;
+    task.index = index;
+    return task;
+  },
+};
